+++ Rebus.StructureMap/MessageHandlerConvention.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Rebus.Handlers;
 using StructureMap.Graph;
 using StructureMap.Graph.Scanning;
@@ -11,29 +8,37 @@ using StructureMap.TypeRules;
 
 namespace Rebus.StructureMap
 {
+    /// <summary>
+    /// StructureMap <see cref="IRegistrationConvention"/> that registers found Rebus handlers in the container
+    /// </summary>
     public class MessageHandlerConvention : IRegistrationConvention
     {
+        /// <summary>
+        /// Registers found Rebus handler types in the container
+        /// </summary>
         public void ScanTypes(TypeSet types, global::StructureMap.Registry registry)
         {
             var messageHandlers = types.FindTypes(TypeClassification.Concretes)
                 .Where(t => t.CanBeCastTo(typeof(IHandleMessages)));
-            foreach (var t in messageHandlers)
+
+            foreach (var handlerType in messageHandlers)
             {
-                var handlers = t.GetInterfaces().Where(IsHandler).ToList();
-                if (handlers.Any())
+                var handlerInterfaces = handlerType.GetInterfaces().Where(IsHandler).ToList();
+
+                foreach (var handlerInterface in handlerInterfaces)
                 {
-                    foreach (var h in handlers)
-                    {
-                        registry.For(h).Use(t).LifecycleIs<UniquePerRequestLifecycle>();
-                    }
+                    registry
+                        .For(handlerInterface)
+                        .Use(handlerType)
+                        .LifecycleIs<UniquePerRequestLifecycle>();
                 }
             }
         }
 
-        private static bool IsHandler(Type i)
+        static bool IsHandler(Type type)
         {
-            return i.IsGenericType &&
-                   (i.GetGenericTypeDefinition() == typeof(IHandleMessages<>));
+            return type.IsGenericType
+                   && type.GetGenericTypeDefinition() == typeof(IHandleMessages<>);
         }
     }
 }